+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/request/OverlayRequest.java
@@ -10,6 +10,10 @@ import javax.validation.constraints.NotEmpty;
 public class OverlayRequest {
     @NotEmpty
     private String label;
+
+    @NotEmpty
+    private String layerName;
+
     @NotEmpty
     private String url;
 }