+++ s4e-backend/src/main/java/pl/cyfronet/s4e/service/OverlayService.java
@@ -5,12 +5,10 @@ import lombok.val;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.cyfronet.s4e.bean.*;
-import pl.cyfronet.s4e.controller.response.AppUserResponse;
 import pl.cyfronet.s4e.controller.response.OverlayResponse;
-import pl.cyfronet.s4e.data.repository.InstitutionRepository;
 import pl.cyfronet.s4e.data.repository.WMSOverlayRepository;
+import pl.cyfronet.s4e.properties.GeoServerProperties;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -19,82 +17,94 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class OverlayService {
     private final WMSOverlayRepository wmsOverlayRepository;
-    private final InstitutionRepository institutionRepository;
+    private final GeoServerProperties geoServerProperties;
 
-    @Transactional(readOnly = true)
-    public List<OverlayResponse> findAllGlobalBy(String geoserverUrl) {
-        return allWmsToResponses(
-                wmsOverlayRepository
-                        .findAllByOwnerType(OverlayOwner.GLOBAL, WMSOverlay.class),
-                geoserverUrl
-        );
+    public List<OverlayResponse> findAllGlobalByUser(AppUser user) {
+        val overlays = wmsOverlayRepository
+                .findAllByOwnerType(OverlayOwner.GLOBAL);
+        val nonActiveIds = user.getPreferences().getNonVisibleOverlays();
+        return mapToOverlaysResponses(overlays, nonActiveIds);
     }
 
-    public List<OverlayResponse> findAllInstitutional(AppUserResponse user, String geoserverUrl) {
-        return user.getRoles().stream()
-                .filter(userRoleResponse -> userRoleResponse.getRole().equals(AppRole.GROUP_MEMBER))
-                .map(userRoleResponse -> {
-                    val institution = institutionRepository.findBySlug(userRoleResponse.getInstitutionSlug(), Institution.class).get();
-                    return allWmsToResponses(
-                            wmsOverlayRepository
-                                    .findAllByOwnerTypeAndInstitutionId(OverlayOwner.INSTITUTIONAL, institution.getId(), WMSOverlay.class),
-                            geoserverUrl
-                    );
-                })
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
+    public List<OverlayResponse> findAllGlobal() {
+        val overlays = wmsOverlayRepository
+                .findAllByOwnerType(OverlayOwner.GLOBAL);
+        List<Long> nonActiveIds = List.of();
+        return mapToOverlaysResponses(overlays, nonActiveIds);
     }
 
-    public List<OverlayResponse> findAllPersonalBy(AppUser appUser, String geoserverUrl) {
-        return allWmsToResponses(
-                wmsOverlayRepository
-                        .findAllByOwnerTypeAndAppUserId(OverlayOwner.PERSONAL, appUser.getId(), WMSOverlay.class),
-                geoserverUrl
-        );
+    public List<OverlayResponse> findAllInstitutionalByUser(AppUser user) {
+        val overlays = wmsOverlayRepository
+                .findAllInstitutional(
+                        user.getId(),
+                        AppRole.GROUP_MEMBER,
+                        OverlayOwner.INSTITUTIONAL
+                );
+        val nonActiveIds = user.getPreferences().getNonVisibleOverlays();
+        return mapToOverlaysResponses(overlays, nonActiveIds);
     }
 
-    public <T> Optional<T> findByOwnerTypeAndIdAndInstitutionId(OverlayOwner owner, Long id, Long institutionId, Class<T> projection) {
-        return wmsOverlayRepository.findByOwnerTypeAndIdAndInstitutionId(owner, id, institutionId, projection);
+    public List<OverlayResponse> findAllPersonalByUser(AppUser appUser) {
+        val overlays = wmsOverlayRepository
+                .findAllPersonal(appUser.getId(), OverlayOwner.PERSONAL);
+        val nonActiveIds = appUser.getPreferences().getNonVisibleOverlays();
+        return mapToOverlaysResponses(overlays, nonActiveIds);
     }
 
-    public <T> Optional<T> findByOwnerTypeAndIdAndUserId(OverlayOwner owner, Long id, AppUser appUser, Class<T> projection) {
-        return wmsOverlayRepository.findByOwnerTypeAndIdAndAppUserId(owner, id, appUser.getId(), projection);
+    public <T> Optional<T> findByIdAndOwnerTypeAndInstitutionId(
+            Long id,
+            OverlayOwner owner,
+            Institution institution,
+            Class<T> projection
+    ) {
+        return wmsOverlayRepository
+                .findByIdAndOwnerTypeAndInstitutionId(id, owner, institution.getId(), projection);
     }
 
-    public <T> Optional<T> findByOwnerTypeAndId(OverlayOwner owner, Long id, Class<T> projection) {
-        return wmsOverlayRepository.findByOwnerTypeAndId(owner, id, projection);
+    public <T> Optional<T> findByIdAndAppUserIdAndOwnerType(
+            Long id,
+            AppUser appUser,
+            OverlayOwner owner,
+            Class<T> projection
+    ) {
+        return wmsOverlayRepository
+                .findByIdAndAppUserIdAndOwnerType(id, appUser.getId(), owner, projection);
     }
 
     @Transactional
-    public OverlayResponse save(WMSOverlay overlay) {
-        return wmsToResponse(wmsOverlayRepository.save(overlay))
-                .url(overlay.getUrl())
-                .build();
+    public void deleteByIdAndOwnerType(Long id, OverlayOwner owner) {
+        wmsOverlayRepository.deleteByIdAndOwnerType(id, owner);
     }
+
     @Transactional
     public void delete(WMSOverlay overlay) {
         wmsOverlayRepository.delete(overlay);
     }
 
-    private List<OverlayResponse> allWmsToResponses(List<WMSOverlay> overlays, String geoserverUrl) {
-        return overlays
-                .stream()
-                .map(wmsOverlay -> wmsToResponse(wmsOverlay)
-                        .url(
-                                wmsOverlay.getUrl().isEmpty()
-                                        ? geoserverUrl + wmsOverlay.getLabel()
-                                        : wmsOverlay.getUrl()
-                        ).build()
-                )
-                .collect(Collectors.toList());
+    @Transactional
+    public OverlayResponse save(WMSOverlay overlay) {
+        return toOverlayResponse(wmsOverlayRepository.save(overlay), List.of());
     }
 
-    private OverlayResponse.OverlayResponseBuilder wmsToResponse(WMSOverlay overlay) {
+    private List<OverlayResponse> mapToOverlaysResponses(List<WMSOverlay> overlays, List<Long> nonActiveIds) {
+        return overlays.stream()
+            .map(overlay -> toOverlayResponse(overlay, nonActiveIds))
+            .collect(Collectors.toList());
+    }
+
+    private OverlayResponse toOverlayResponse(WMSOverlay overlay, List<Long> nonActiveIds) {
         return OverlayResponse.builder()
                 .id(overlay.getId())
-                .label(overlay.getLabel())
-                .visible(true)
+                .layerName(overlay.getLayerName())
+                .url(getOverlayUrlBy(overlay.getUrl()))
                 .createdAt(overlay.getCreatedAt())
-                .ownerType(overlay.getOwnerType().toString());
+                .ownerType(overlay.getOwnerType().toString())
+                .label(overlay.getLabel())
+                .visible(!nonActiveIds.contains(overlay.getId()))
+                .build();
+    }
+
+    private String getOverlayUrlBy(String actualUrl) {
+        return actualUrl.isEmpty() ? geoServerProperties.getOutsideBaseUrl() : actualUrl;
     }
 }