+++ s4e-backend/src/test/java/pl/cyfronet/s4e/OverlayHelper.java
@@ -20,6 +20,7 @@ public class OverlayHelper {
         val url = nextUnique(OverlayHelper.url);
         return OverlayRequest.builder()
                 .label(label)
+                .layerName(label)
                 .url(url);
     }
 
@@ -28,6 +29,7 @@ public class OverlayHelper {
         val url = nextUnique(OverlayHelper.url);
         return WMSOverlay.builder()
                 .label(label)
+                .layerName(label)
                 .url(url);
     }